pyproject.toml
[project]
name = "hyperevm-cex-dex-estimation"
version = "0.1.0"
description = ""
authors = [
    {name = "kosunghun317",email = "kosunghun317@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "ipykernel (>=6.30.1,<7.0.0)",
    "ruff (>=0.12.11,<0.13.0)",
    "polars (>=1.33.0,<2.0.0)",
    "tardis-client (>=1.3.10,<2.0.0)",
    "mamushi (>=0.0.8,<0.0.9)",
    "moccasin (>=0.4.2,<0.5.0)",
    "web3 (>=7.13.0,<8.0.0)",
    "dune-client (>=1.7.10,<2.0.0)",
    "tardis-dev (>=2.1.2,<3.0.0)",
    "hyperliquid-python-sdk (>=0.18.0,<0.19.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "pyarrow (>=21.0.0,<22.0.0)"
]

[tool.poetry]
packages = [{include = "hyperevm_cex_dex_estimation", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
